refactor: seperate service function into file
