--- server/src/services/index.ts
@@ -1,8 +1,0 @@
-export { getEventById, getUserEventsByUserId } from './events';
-export { getUserById, getUserByGoogleId, setUser, setUserInfo } from './users';
-export {
-  getUserTicketsByUserId,
-  getUserTicketsByTicketId,
-  toggleUserAttendance,
-  deleteUserTicketById,
-} from './userTickets';
