+++ server/src/services/users/getUserById.ts
@@ -1,0 +1,7 @@
+import { User } from 'models';
+import { WhereOptions } from 'sequelize';
+
+export default async (id: number): Promise<User | null> => {
+  const where: WhereOptions = { id };
+  return await User.findOne({ where });
+};
